requirements.txt
streamlit
google-generativeai
supabase
pandas
Pillow